requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-environ==0.9.0
django-import-export==2.8.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.2.0
waitress==2.1.2
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
